저장용저장용

+++ AllThatTip/src/main/java/com/kh/AllThatTrip/member/model/service/MemberServiceImpl.java
@@ -28,28 +28,23 @@ public class MemberServiceImpl implements MemberService {
 	public Member login(Member member) {
 		return mapper.login(member);
 	}
-	
 	@Override
 	public int join(Member member) {
 		validator.validateJoinMember(member);
 		member.setUserPwd(passwordEncoder.encode(member.getUserPwd()));
 		return mapper.join(member);
 	}
-
 	@Override
 	public int checkId(String userId) {
 		return mapper.checkId(userId); 
 	}
-
 	@Override
 	public void memberUpdate(Member member, HttpSession session) {
-//		validator.validateLength(member);
 		validator.validateMail(member);
 		member.setUserPwd(passwordEncoder.encode(member.getUserPwd()));
 		mapper.memberUpdate(member);
 		session.setAttribute("loginUser", mapper.login(member));
 	}
-
 	@Override
 	public void memberDelete(String userPwd, HttpSession session) {
 		Member loginUser = (Member)session.getAttribute("loginUser");
@@ -62,12 +57,10 @@ public class MemberServiceImpl implements MemberService {
 		session.setAttribute("deleteSuccess", userInfo);
 		session.removeAttribute("loginUser");
 	}
-
 	@Override
 	public String findId(String userName) {
 		return mapper.findId(userName);
 	}
-
 	@Override
 	public void findPwd(Member member) {
 		validator.validateMail(member);
@@ -79,36 +72,25 @@ public class MemberServiceImpl implements MemberService {
 				mapper.memberUpdate(result);
 			}
 	}
-	
 	@Override
 	public void increaseLoginCount(Member member) {
 		mapper.increaseLoginCount(member);
 	}
-	
 	@Override
 	public void rollbackCount(Member member) {
 		mapper.rollbackCount(member);
 	}
-	
 	@Override
 	public Member countCheck(Member member) {
-		
 		Member result = mapper.countCheck(member);
-		
-		if(result == null) {
-			throw new LoginFailedException("없는 ID입니다.");
-		}
-		
+		if(result == null) { throw new LoginFailedException("없는 ID입니다."); }
 		return result;
 	}
-	
 	@Override
 	public void loginFullCount(Member member) {
 		mapper.loginFullCount(member);
 	}
-	
 	/* 찜, 예약, 장바구니 관련 */
-
 	@Override
 	public void findRsv(Member member, HttpSession session) {
 		session.removeAttribute("findRsvResult");
@@ -117,20 +99,13 @@ public class MemberServiceImpl implements MemberService {
 			session.setAttribute("findRsvResult", result);
 		}
 	}
-	
-	// CART GET요청 시
 	@Override
 	public void findCart(HttpSession session) {
-		if(session.getAttribute("loginUser") == null) {
-			// 로그인 되어있지 않다면 장바구니 초기화
-			mapper.deleteCart();
-		}
+		if(session.getAttribute("loginUser") == null) { mapper.deleteCart(); }
 		ArrayList cartResult = mapper.findCart();
 		log.info("카트조회결과{}", cartResult);
 		session.setAttribute("cartResult", cartResult);
 	}
-	
-	// CART POST 요청 시 (장바구니 등록 경우)
 	@Override
 	public void findRoom(Room room, HttpSession session) {
 		if(session.getAttribute("loginUser") == null) {
@@ -141,41 +116,4 @@ public class MemberServiceImpl implements MemberService {
 		ArrayList cartResult = mapper.findCart();
 		session.setAttribute("cartResult", cartResult);
 	}
-
-
-
-
-	
-
-
-
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
 }
